requirements.txt
numpy>=1.21.0
torch>=2.0.0
pandas>=1.3.0
pytest>=7.0.0
pytest-cov>=4.0.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
argparse>=1.4.0
flake8>=3.9.0
pydantic>=2.0
optuna>=3.0
pyyaml>=6.0
joblib>=1.2.0
